Move monitor threads and friends to own package
This will avoid collisions when both bootstrap 2.1 and above is present
