+++ nexus/nexus-bootstrap/src/main/java/org/sonatype/nexus/bootstrap/monitor/commands/StopApplicationCommand.java
@@ -10,34 +10,48 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.bootstrap.commands;
+package org.sonatype.nexus.bootstrap.monitor.commands;
 
-import org.sonatype.nexus.bootstrap.CommandMonitorThread;
+import org.sonatype.nexus.bootstrap.monitor.CommandMonitorThread;
+import org.sonatype.nexus.bootstrap.Launcher;
 import org.sonatype.nexus.bootstrap.log.LogProxy;
 
 /**
- * Stops command monitor.
+ * Stop launcher.
  *
  * @since 2.2
  */
-public class StopMonitorCommand
+public class StopApplicationCommand
     implements CommandMonitorThread.Command
 {
 
-    private static final LogProxy log = LogProxy.getLogger( StopMonitorCommand.class );
+    private static final LogProxy log = LogProxy.getLogger( StopApplicationCommand.class );
 
-    public static final String STOP_MONITOR_COMMAND = "STOP_MONITOR";
+    public static final String STOP_APPLICATION_COMMAND = "STOP";
+
+    private final Launcher launcher;
+
+    public StopApplicationCommand( final Launcher launcher )
+    {
+        if ( launcher == null )
+        {
+            throw new NullPointerException();
+        }
+        this.launcher = launcher;
+    }
 
     @Override
     public String getId()
     {
-        return STOP_MONITOR_COMMAND;
+        return STOP_APPLICATION_COMMAND;
     }
 
     @Override
     public boolean execute()
     {
-        log.info( "Requesting monitor stop" );
+        log.info( "Requesting application stop" );
+        launcher.commandStop();
+
         return true;
     }
